pyproject.toml
[project]
name = "philly"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fire[cli]>=0.7.0",
    "geojson>=3.2.0",
    "geoparquet>=0.0.3",
    "gtfs-realtime-bindings>=1.0.0",
    "geopandas>=1.1.0",
    "httpx>=0.28.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>0",
    "basedpyright>0",
    "poethepoet>=0.37.0",
]
test = ["pytest>0", "pytest-asyncio>0"]
notebook = [
    "catppuccin-jupyterlab>=0.2.4",
    "ipykernel>0",
    "jupyterlab>0",
    "jupyterlab-vim>=4.1.4",
]

[project.scripts]
phl = "philly.__main__:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"philly" = ["datasets/*.yaml"]

[tool.poe.tasks]
fmt = "uv run ruff format"
lint = "uv run ruff check --fix"
type-check = "uv run basedpyright --level error"
test = "uv run pytest tests"

all = ["fmt", "lint", "type-check"]
